requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Document processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0

# Utilities
tiktoken==0.5.2
tqdm==4.66.1
streamlit==1.29.0

# Optional: For better embeddings
openai>=1.10.0,<2.0.0 